refactor: coreモジュールとutilsパッケージを作成し、main.py内の処理を切り分ける  #14
